JavaScriptの実装
