--- app/javascript/items/new.js
@@ -1,15 +1,0 @@
-function calc() {
-  const form = document.getElementById("item-price");
-  const tax = document.getElementById("add-tax-price");
-  const profit = document.getElementById("profit");
-
-  form.addEventListener('input', () => {
-    let price = parseInt(form.value, 10);
-    let taxRes = parseInt(price / 10);
-    let profitRes = parseInt(price / 10 * 9);
-    tax.innerHTML = taxRes.toLocaleString();
-    profit.innerHTML = profitRes.toLocaleString();
-  });
-}
-
-window.addEventListener('load', calc);
