+++ app/javascript/packs/application.js
@@ -7,7 +7,7 @@ require("@rails/ujs").start()
 require("turbolinks").start()
 require("@rails/activestorage").start()
 require("channels")
-require("../items/new")
+require('../new')
 
 
 // Uncomment to copy all static images under ../images to the output folder and reference
